address a model file in a width that can hold it
A 7B checkpoint's later tensors sit past four gigabytes, and usize is
thirty-two bits on wasm32 — the one target where a model that large is
the entire point. A 447 MB tensor at offset 4.24 GB summed to 4.68 GB
and wrapped, so the reader reported a byte range whose end came before
its own start and refused to serve it. Nothing above four gigabytes was
reachable before per-tensor streaming, which is why this waited.

File offsets are u64 now, from the header through the window to the
mount's cursor; only indices INTO a held buffer stay usize, and those
cannot exceed the address space that holds them. A test states the
requirement in the numbers that failed.

The window is also moved into the reader and back rather than copied
per tensor. Copying cost the whole model in memcpy and, at the moment
the widest tensor was resident, kept two copies of it live — which is
most of the difference between fitting a tab and not.

In a tab: a 4.68 GB model mounts in 25 s and answers six questions,
where before it could not be addressed at all. Linear memory peaks at
2.16 GB, above the 1.5 GB written down beforehand; the window is not
what holds it — the packed weights are repacked through a host buffer
on their way to the device, so the widest tensor is briefly present
twice. Named, not fixed.

+++ engine/core/combs-wasm/src/lib.rs
@@ -291,7 +291,7 @@ pub async fn combs_model_open(
         let state = Mount::Streaming {
             config_json,
             mount: combs_runtime::stream_mount::StreamMount::new(
-                expected_len as usize,
+                expected_len as u64,
                 combs_core::init_device(),
             ),
         };
